requirements.txt
pydantic_settings
sqlalchemy
fastapi
pwdlib[argon2]
uvicorn
asyncpg
alembic
python-multipart
pyjwt